+++ tools/chessops/tool.js
@@ -17,11 +17,11 @@
     intl = {
       'en-US': {
         'options.study': 'Study',
-        'options.chessOps': 'chessops support'
+        'options.chessOps': 'Chessops support'
       },
       'ro-RO': {
         'options.study': 'Studiu',
-        'options.chessOps': 'suport chessops'
+        'options.chessOps': 'Suport pentru chessops'
       }
     }
 
@@ -29,6 +29,7 @@
       const parent = this.lichessTools;
       const value = parent.currentOptions.getValue('chessOps');
       if (!value) {
+        parent.chessops=null;
         return;
       }
       parent.chessops = (function () {