+++ tools/downloadAllStudies/tool.js
@@ -42,7 +42,7 @@
       if (!m) return;
       const userId=m[1].toLowerCase();
       $('<a class="lichessTools-downloadAllStudies">')
-        .attr('href','https://lichess.org/study/by/'+userId+'/export.pgn')
+        .attr('href','/study/by/'+userId+'/export.pgn')
         .attr('title',trans.noarg('downloadAllTitle'))
         .text(trans.noarg('downloadAllText'))
         .appendTo('div.box__top');