+++ tools/extraChart/tool.js
@@ -9,7 +9,7 @@
         category: 'analysis',
         type:'multiple',
         possibleValues: ['material','principled','tension','potential','brilliant','smooth','gauge'],
-        defaultValue: 'material,principled,tension,smooth,gauge',
+        defaultValue: 'material,principled,tension,smooth,gauge,brilliant',
         advanced: true
       },
       {
@@ -39,7 +39,8 @@
         'potentialLineTitle': 'Max potential',
         'goodMovesText':'good/brilliant/interesting moves',
         'goodMovesTitle':'LiChess Tools - good/brilliant/interesting moves',
-        'merryChristmas':'Merry Christmas from LiChess Tools!'
+        'merryChristmas':'Merry Christmas from LiChess Tools!',
+        'options.christmas': 'Show Christmas lights on chart on the 25th of December'
       },
       'ro-RO':{
         'options.analysis': 'Analiz\u0103',
@@ -709,8 +710,8 @@
       const initActiveElements=chart.getActiveElements()?.map(e=>{ return {datasetIndex:e.datasetIndex, index:e.index}; });
       const colors=['red','green','blue','orange','yellow','magenta','cyan','white'];
       for (let i=0; i<30; i++) {
-        const elements=dataset.data.filter(d=>Math.random()<0.3).map(d=>{ return { datasetIndex:0, index:d.x-1 }; });
-        const color=colors[Math.round(Math.random()*colors.length)];
+        const elements=dataset.data.filter(d=>parent.random()<0.3).map(d=>{ return { datasetIndex:0, index:d.x-1 }; });
+        const color=colors[Math.round(parent.random()*colors.length)];
         dataset.hoverBackgroundColor=color;
         chart.setActiveElements(elements);
         chart.update('none');