+++ tools/fixAbortController/tool.js
@@ -26,7 +26,9 @@
       const parent=this.lichessTools;
       const value=parent.currentOptions.getValue('fixAbortController');
       this.logOption('Fix AbortController in debug mode', value);
-      parent.global.AbortController.prototype.abort=parent.wrapFunction(AbortController.prototype.abort,{ 
+      parent.global.AbortController.prototype.abort=parent.unwrapFunction(parent.global.AbortController.prototype.abort,'fixAbortController');
+      if (!value) return;
+      parent.global.AbortController.prototype.abort=parent.wrapFunction(parent.global.AbortController.prototype.abort,{ 
         id:'fixAbortController',
         before: ($this,...args) => $this.signal?.aborted
       });