+++ tools/stockfish/tool.js
@@ -15,11 +15,11 @@
     intl = {
       'en-US': {
         'options.general': 'General',
-        'options.stockfish': 'stockfish'
+        'options.stockfish': 'Stockfish'
       },
       'ro-RO': {
         'options.general': 'General',
-        'options.stockfish': 'stockfish'
+        'options.stockfish': 'Stockfish'
       }
     }
 
@@ -27,6 +27,8 @@
       const parent = this.lichessTools;
       const value = parent.currentOptions.getValue('stockfish');
       if (!value) {
+        parent.stockfish?.destroy();
+        parent.stockfish=null;
         return;
       }
       const sf=new Stockfish(parent);